Add asynchronous request for posts on index
Following the advice on how to set up fetch requests in a react app
here: https://www.robinwieruch.de/react-fetching-data#how-to-fetch-data-in-react

I managed to overload the componentDidMount method of my app, so that
the fetch request is made once the App component has rendered once. This
means we're only making one fetch request, instead of continuously
making them inside the render method.