Add api create method
With this I should be able to submit post requests with foreign key ids,
and they will be converted to the actual objects, so the target model
can be created.

Also need to hook up a form to the post request, and then add feedback
on the create method's return (success - render the new post with the
others; failure - render an error message).