Add create post functionality
Tried to work out how to get the csrf token using cookies, but it looks
like it requires installing third party plugins, which I don't really
want to get in to right now. My workaround is the add the csrf token the
layout template, and then simply reference that in the POST requests.

Had to modify the create function slightly, because the data is no
longer part of the request.POST, but only present in the request.body.

Also, the user comes as part of the request, and it was simpler to add
the user separately to the other params. This is why I added the
**kwargs param to the create_from_post, so that even on models that
don't require the user, the argument passed in from the create pathway
can be safely ignored.