Add mechanism to display like count on posts
Currently static, so the counter starts to break above 99 likes (the
text doesn't fit in the container) so this will have to be fixed at some
point.

Also need to do a general colour styling at some point, as it's just
black and white at the moment. Quite like the react cyan on charcoal
vibe while I've been researching for this project, so I might go for
that.