Bugfix on seach api with null request model
Python's dict.get function will only fill in the default value if the
key is not present. If the key is present and set to null/None, however,
search api would attempt to get a model of name 'None' instead of a
string, which throws a ValueError instead of a LookupError.

Also added quick todo because I'm not doing any handling on error
responses or catching any other errors