Sort posts on profile in reverse chronological order
To implement this I had to add functionality to api multi link field
syntax so that multi-link fields can now be either a list of fields as
before, or a special objects with keys 'fields' for fields, and 'order'
to specify the order the multi-link models should be returned.

Also added tests to verify this functionality.