Refactor navigation bar click functions to react methods
Planning to update the state based on clicking the navigation bar items
so for now I've refactored them to regular divs and bound a react method
to each of their onclick event handlers.

Viewing the user's own profile was the easiest to implement so I've
added that in already. The others will require building the
functionality and refactoring the render method.

Left the login, register and logout links as regular django pathways
because I don't want to mess around with authentication for now.