Create generalised api search method in React app
This will stop me having to copy and paste a bunch of boiler plate
search api fetch requests later for the other pages.

I also slightly refactored the parameter sanitization for filters from
the view function, into the serialization function itself. This is
because in python if a key is available, but None, it will not resolve
to the default from the dict.get(). This would be a problem if I'm using
a generalised search method, because I would have to pass in a null
filter any time no filtering was happeing (or do some fiddly if-else).
Plus it's better for the serialize function to be responsible for it's
own parameters rather than an external value. It'll be better this way,
trust me. I don't know why I've written an essay to justify a tiny
change on a private repo... but here we are.