Bugfix select all on lower level multi linked field
The select all flag wasn't working when trying to get all fields from a
linked field from a lower level. Fixed now so that it resolves
correctly.
While testing this, I also noticed foreign key fields were not being
evaluated correctly because it was using the model class, not the model
instance - so values were coming back as None.