Add method for react app to determine current user
React app now has a way to dynamically retrieve current user info
without having to refresh the page. It's a bit redundant to do it this
way, because we already have access to the user object when serving the
html template - but I don't really like pulling variables from the html
to use within react app state, feels a bit hacky.

Anyway, I'll fix this if I run into issues, but for now this will work.