Refactor view profile into two subqueries
Decided it would be easier to use the existing pagination logic provided
by the search view, rather than try to re-implement it inside multi-link
fields because that would mean having a data structure like
user.leaders.data.<fields> instead of user.leaders.<fields> which seems
a bit nicer (even though I'm not actually using it now...)

While doing this, noticed a bug in the filter syntax where if a user is
not following anyone the regex fails. Been meaning to refactor the
filter syntax anyway, so this is a good reason to.