Fix Anonymous user redirect to login
If a user is not logged in they can attempt to create a post or add a
like, but they will get a redirect response. This fix redirects the page
to the login screen.

Ideally after logging in the user would be taken back to where they were
and have their action completed.

Also got a bug where users can't get back to 'All Posts' from the login
screen because the React app is not currently available.